HW7.2
