+++ app/src/main/java/ru/oliverhd/weather/WeatherDetailFragment.java
@@ -8,12 +8,28 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import android.os.Handler;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import com.google.gson.Gson;
+
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.stream.Collectors;
+
+import javax.net.ssl.HttpsURLConnection;
+
+import ru.oliverhd.weather.model.WeatherRequest;
+
 public class WeatherDetailFragment extends Fragment implements Constants {
+
+    private static final String TAG = "Weather";
 
     public static WeatherDetailFragment create(Parcel parcel) {
         WeatherDetailFragment fragment = new WeatherDetailFragment();
@@ -38,13 +54,49 @@ public class WeatherDetailFragment extends Fragment implements Constants {
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        TextView textView = view.findViewById(R.id.weather_info);
+        final TextView textView = view.findViewById(R.id.weather_info);
         TextView city = view.findViewById(R.id.city);
         Parcel parcel = getParcel();
         city.setText(parcel.getCityName());
+        try {
+            final String url = String.format("https://api.openweathermap.org/data/2.5/weather?q=%s&appid=%s", city.getText(), BuildConfig.WEATHER_API_KEY);
+            final URL uri = new URL(url);
+            final Handler handler = new Handler();
+            new Thread(new Runnable() {
+                public void run() {
+                    HttpsURLConnection urlConnection = null;
+                    try {
+                        urlConnection = (HttpsURLConnection) uri.openConnection();
+                        urlConnection.setRequestMethod("GET");
+                        urlConnection.setReadTimeout(10000);
 
-        WeatherHandler weatherHandler = new WeatherHandler(parcel.getCityName());
-        textView.setText(weatherHandler.getTemperature());
+                        BufferedReader in = new BufferedReader(new InputStreamReader(urlConnection.getInputStream()));
+                        String result = getLines(in);
+
+                        Gson gson = new Gson();
+                        final WeatherRequest weatherRequest = gson.fromJson(result, WeatherRequest.class);
+                        handler.post(new Runnable() {
+                            @Override
+                            public void run() {
+                                textView.setText(Float.toString((weatherRequest.getMain().getTemp() - 273)));
+                            }
+                        });
+
+
+                    } catch (Exception e) {
+                        Log.e(TAG, "Fail connection", e);
+                        e.printStackTrace();
+                    } finally {
+                        if (urlConnection != null) {
+                            urlConnection.disconnect();
+                        }
+                    }
+                }
+            }).start();
+        } catch (MalformedURLException e) {
+            Log.e(TAG, "Fail URI", e);
+            e.printStackTrace();
+        }
 
         String[] data = getResources().getStringArray(R.array.time);
 
@@ -57,5 +109,9 @@ public class WeatherDetailFragment extends Fragment implements Constants {
 
         DetailAdapter adapter = new DetailAdapter(data);
         recyclerView.setAdapter(adapter);
+    }
+
+    private String getLines(BufferedReader in) {
+        return in.lines().collect(Collectors.joining("\n"));
     }
 }
